+++ src/main/kotlin/com/bugsnag/android/gradle/AndroidManifestInfoReceiver.kt
@@ -1,9 +1,13 @@
 package com.bugsnag.android.gradle
 
-import org.gradle.api.provider.Property
+import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.tasks.Input
 
 interface AndroidManifestInfoReceiver {
     @get:Input
-    val manifestInfo: Property<AndroidManifestInfo>
+    val manifestInfo: RegularFileProperty
+}
+
+internal fun AndroidManifestInfoReceiver.parseManifestInfo(): AndroidManifestInfo {
+    return AndroidManifestInfo.read(manifestInfo.asFile.get())
 }