+++ src/main/kotlin/com/bugsnag/android/gradle/BugsnagReleasesTask.kt
@@ -19,6 +19,7 @@ import org.gradle.api.provider.MapProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.ProviderFactory
 import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.Optional
@@ -55,8 +56,8 @@ open class BugsnagReleasesTask @Inject constructor(
         description = "Assembles information about the build that will be sent to the releases API"
     }
 
-    @get:Input
-    override val manifestInfo: Property<AndroidManifestInfo> = objects.property()
+    @get:InputFile
+    override val manifestInfo: RegularFileProperty = objects.fileProperty()
 
     @get:Internal
     internal val uploadRequestClient: Property<UploadRequestClient> = objects.property()
@@ -141,7 +142,7 @@ open class BugsnagReleasesTask @Inject constructor(
 
     @TaskAction
     fun fetchReleaseInfo() {
-        val manifestInfo = manifestInfo.get()
+        val manifestInfo = parseManifestInfo()
         val payload = generateJsonPayload(manifestInfo)
 
         val response =